+++ ChangeLog.md
@@ -1,3 +1,10 @@
 # Changelog for cobot
 
 ## Unreleased changes
+
+## [0.1.1.0] - 2019-06-17
+### Added
+- Typeclass `IsGap`.
+- Now you can align two sequences using different gaps for each one of them.
+### Changed
+- Removed `affine` function from `SequenceAlignment` typeclass.